NN: merge r23217 from branches/jag/2.2-sep19
git-svn-id: http://svn.zentyal.org/zentyal/trunk@23263 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ main/bwmonitor/schemas/sql/bwmonitor_usage.sql
@@ -4,18 +4,18 @@ CREATE TABLE bwmonitor_usage (
     interface     VARCHAR(30),
 
     /* internal traffic */
-    intTotalRecv  INTEGER,
-    intTotalSent  INTEGER,
-    intTCP        INTEGER,
-    intUDP        INTEGER,
-    intICMP       INTEGER,
+    intTotalRecv  BIGINT,
+    intTotalSent  BIGINT,
+    intTCP        BIGINT,
+    intUDP        BIGINT,
+    intICMP       BIGINT,
 
     /* external traffic */
-    extTotalRecv  INTEGER,
-    extTotalSent  INTEGER,
-    extTCP        INTEGER,
-    extUDP        INTEGER,
-    extICMP       INTEGER,
+    extTotalRecv  BIGINT,
+    extTotalSent  BIGINT,
+    extTCP        BIGINT,
+    extUDP        BIGINT,
+    extICMP       BIGINT,
 
     timestamp TIMESTAMP);
